Improve reliability of QR point match
The polygon given by pyzbar has the corners in an arbitrary order. It's
difficult to sort too. If there is a small angle, the top left corner
could be to the left or right of the bottom left corner.

The most reliable method is to try all the permutations and to take the
one with the smallest distance between the points of the straight
bounding box and the perspective bounding box.